// src/api/mirakc.d.ts
export interface components {
    schemas: {
// ...
        MirakurunProgram: {
            audio?: {
                /** Format: int32 */
                componentType: number;
                isMain: boolean;
                langs: string[];
                /** Format: int32 */
                samplingRate: number;
            } | null;
            audios?: {
                /** Format: int32 */
                componentType: number;
                isMain: boolean;
                langs: string[];
                /** Format: int32 */
                samplingRate: number;
            }[];
            description?: string | null;
            /** Format: int64 */
            duration: number;
            /** Format: int32 */
            eventId: number;
            extended?: Record<string, never>;
            genres?: {
                /** Format: int32 */
                lv1: number;
                /** Format: int32 */
                lv2: number;
                /** Format: int32 */
                un1: number;
                /** Format: int32 */
                un2: number;
            }[] | null;
            /** Format: int64 */
            id: number;
            isFree: boolean;
            name?: string | null;
            /** Format: int32 */
            networkId: number;
            relatedItems?: {
                /** Format: int32 */
                eventId: number;
                /** Format: int32 */
                networkId?: number | null;
                /** Format: int32 */
                serviceId: number;
                type: string;
            }[];
            series?: {
                /** Format: int32 */
                episode: number;
                /** Format: int64 */
                expireAt: number;
                /** Format: int32 */
                id: number;
                /** Format: int32 */
                lastEpisode: number;
                name: string;
                /** Format: int32 */
                pattern: number;
                /** Format: int32 */
                repeat: number;
            } | null;
            /** Format: int32 */
            serviceId: number;
            /** Format: int64 */
            startAt: number;
            video?: {
                /** Format: int32 */
                componentType: number;
                resolution?: string | null;
                /** Format: int32 */
                streamContent: number;
                type?: string | null;
            } | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}